Conflicts:
	Project Loyal/bin/.gitignore
	Project Loyal/bin/loyal/Loyal.class
	Project Loyal/bin/loyal/level/Level.class
	Project Loyal/src/loyal/Loyal.java
	Project Loyal/src/loyal/entities/MapPlayer.java
	Project Loyal/src/loyal/entities/pointer.java
	Project Loyal/src/loyal/level/Level.java
	Project Loyal/src/loyal/level/LevelGenerator.java

+++ Project Loyal/src/loyal/Battle/Characters/Sword.java
@@ -14,7 +14,7 @@ public class Sword extends ModifyingItem{
 
 	@Override
 	public void unSetModified() {
-		int temp = innerState.getStat(Str);
+		double temp = innerState.getStat(Str);
 		if (modifiedStateArray[Str.index] > temp)
 			modifiedStateArray[Str.index] = temp;		
 	}
